tidy: skip submodules if not present for non-CI environments
